Advanced-AI-Agent/advanced-agent/pyproject.toml
[project]
name = "advanced-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "firecrawl-py>=4.16.0",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.8",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
